renamed files as snack_case
