+++ lib/ui_components/sber_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:effective_flutter_lab/models/Service.dart';
+import 'package:effective_flutter_lab/models/service.dart';
 import 'package:effective_flutter_lab/constants.dart';
 
 class SberCard extends StatelessWidget {
